demos/demo_S1_Small_Model/requirements.txt
# demo_S1_Small_Model — Python 依赖
# 安装：pip install -r requirements.txt

# ---- 核心深度学习 ----
torch>=2.1.0
torchvision>=0.16.0

# ---- HuggingFace 生态 ----
transformers>=4.37.0
accelerate>=0.25.0
datasets>=2.16.0          # 加载 COCO 子集

# ---- 视觉感知指标 ----
lpips>=0.1.4              # LPIPS 感知距离
scikit-image>=0.21.0      # PSNR / SSIM

# ---- 可微分 DCT ----
torch-dct>=0.1.6          # 可微分 DCT 变换（JPEG 模拟）

# ---- 图像处理 ----
Pillow>=10.0.0
numpy>=1.24.0

# ---- 工具 ----
tqdm>=4.66.0
matplotlib>=3.7.0
scipy>=1.11.0             # 高斯滤波

# ---- 可选：BLIP-2 / DeepSeek-VL 专用 ----
# salesforce-lavis         # 若使用 lavis 加载 BLIP-2（可选）
# deepseek_vl              # 若本地安装 DeepSeek-VL 包
# qwen_vl_utils            # Qwen2.5-VL 辅助工具（已随 transformers 安装）

# ---- 开发/测试 ----
pytest>=7.4.0
